:recycle: [#56] Refactor browser client to match webservice
